Incorporate into split
