+++ include/utilities/threshold/param_set.hpp
@@ -1,5 +1,8 @@
 #pragma once
 
+namespace valik
+{
+
 /**
  * @brief A point in the parameter tuning search space.
  * 
@@ -11,7 +14,7 @@ struct param_set
     size_t k;
     size_t t;
 
-    param_set(size_t const kmer_size, size_t const thresh, parameter_space const & space)
+    param_set(size_t const kmer_size, size_t const thresh, param_space const & space)
     {
         std::cout << "Considering k=" << kmer_size << " and t=" << thresh << '\n';
         if (kmer_size < std::get<0>(space.kmer_range) | kmer_size > std::get<1>(space.kmer_range))
@@ -25,3 +28,5 @@ struct param_set
             t = thresh;
     }
 };
+
+}   // namespace valik
