+++ test/api/utilities/CMakeLists.txt
@@ -1,1 +1,2 @@
 add_subdirectory(consolidate)
+add_subdirectory(threshold)
