+++ test/api/utilities/threshold/CMakeLists.txt
@@ -1,0 +1,1 @@
+add_api_test (kmer_attributes_test.cpp)
